requirements.txt
streamlit==1.38.0
qdrant-client==1.9.2
sentence-transformers==3.0.1
rank-bm25==0.2.2
PyPDF2==3.0.1
pandas==2.2.2
python-dotenv==1.0.1
prometheus-client==0.20.0
uvicorn==0.30.1
fastapi==0.111.0
pydantic==2.8.2
nltk==3.9.1
tqdm==4.66.4
requests==2.32.3
pytest==8.3.2
openai>=1.40.0
tiktoken>=0.7.0
